Weathering (& decaying) happens everywhere now
Optimized some repeating code in WeatheringHandler.
Items in shulker boxes now also get weathered (as well as items in shulker boxes in shulker boxes, and items in shulker boxes in shulker boxes in shulker boxes, and items in shulker boxes in shulker boxes in shulker boxes in shulker boxes, and items in shulker boxes in shulker boxes in shulker boxes in shulker boxes in shulker boxes, and items in shulker boxes in shulker boxes in shulker boxes in shulker boxes in shulker boxes in shulker boxes, and items in shulker boxes in shulker boxes in shulker boxes in shulker boxes in shulker boxes in shulker boxes in shulker boxes, and items in shulker boxes in shulker boxes in shulker boxes in shulker boxes in shulker boxes in shulker boxes in shulker boxes in shulker boxes, and items in shulker boxes in shulker boxes in shulker boxes in shulker boxes in shulker boxes in shulker boxes in shulker boxes in shulker boxes in shulker boxes...), and also for any items with minecraft:container component.
Items in furnaces also get weathered, but with some special conditions: see WeatheringHandler comments.
Smelting in furnaces will not stop because of decay mismatch, and when that occurs the result items merge decay values. (attained with mixin)
WeatheringHandler now also sends out packets (DoDecayPacket, ) to clients for certain updates caused by weathering.